refactor: 경로 관련 기능 수정
